Redo GHC.Vacuum.ClosureTypes
The different closure types in GHC are prone to change between versions,
and this module hasn't kept up. In order to prevent incorrect decisions
when computing information about those closures, the safest thing to do
is to have a separate module for every different GHC version.

The top-level 'gen-closuretype.pl' script automatically generates a
particular definition of the 'ClosureType' datatype for a specific
version of GHC. The simple usage is:

  $ ./gen-closuretype.pl > src/GHC/Vacuum/ClosureType/<VER>.hs

where <VER> is representative of __GLASGOW_HASKELL__

You can also pass the name of a particular GHC executable to the
script, i.e. './gen-closuretype.pl ghc-7.0.4 > ...'

The result is conditionally compiled into the library and exported
through GHC.Vacuum.ClosureType for compatibility.

Support for GHC 7.0.* and 7.2.* only as of the moment.

Also removes some old support for IND_OLDGEN and IND_OLDGEN_PERM closure types
in GHC.Vacuum. This is an example of things which may be misrepresented,
as they have not existed for quite a while.

Not sure how far back this axes compatibility.